+++ hardware/X6-1000M/src/libx6adc.h
@@ -60,7 +60,7 @@ EXPORT int set_averager_settings(int, int, int, int, int);
 EXPORT int acquire(int);
 EXPORT int wait_for_acquisition(int, int);
 EXPORT int stop(int);
-EXPORT int transfer_waveform(int, int, long long *, unsigned int);
+EXPORT int transfer_waveform(int, int, double *, unsigned);
 
 EXPORT int set_log(char *);
 int update_log(FILE * pFile);
